refactor fee effects and amounts calculations
remove channels

remove channels
